backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
langchain>=0.0.267
langchain-core>=0.0.12
langchain-community>=0.0.10
langchain-experimental>=0.0.10
langchain-openai>=0.0.2
langchain-google-genai>=0.0.2
neo4j>=5.14.0
pydantic>=2.4.2
python-dotenv>=1.0.0
tiktoken>=0.4.0